refactor(cache): move IQBCache to cache.py
No functional change, just avoiding to implement lots of stuff
inside `__init__.py` because it's not super idiomatic IIUC.
